+++ src/views/ngrest/crud.php
@@ -215,12 +215,12 @@ $this->beginBody();
             </div>
         </div>
         <?php if ($canCreate && $config->getPointer('create')): ?>
-        	<?= $this->render('_crudform', ['type' => '1', 'renderer' => RenderCrud::TYPE_CREATE, 'isInline' => $isInline, 'relationCall' => $relationCall]); ?>
+        	<?= $this->render($this->context->crudFormView, ['type' => '1', 'renderer' => RenderCrud::TYPE_CREATE, 'isInline' => $isInline, 'relationCall' => $relationCall]); ?>
         <?php endif; ?>
         <?php if ($canUpdate && $config->getPointer('update')): ?>
-        	<?= $this->render('_crudform', ['type' => '2', 'renderer' => RenderCrud::TYPE_UPDATE, 'isInline' => $isInline, 'relationCall' => $relationCall]); ?>
+        	<?= $this->render($this->context->crudFormView, ['type' => '2', 'renderer' => RenderCrud::TYPE_UPDATE, 'isInline' => $isInline, 'relationCall' => $relationCall]); ?>
         <?php endif; ?>
-        <?= $this->render('_awform'); ?>
+        <?= $this->render($this->context->awFormView); ?>
     </div>
 </div>
 <?php $this->endBody(); ?>