+++ tests/admin/ngrest/render/RenderCrudTest.php
@@ -3,7 +3,7 @@
 namespace admintests\admin\ngrest\render;
 
 use admintests\AdminTestCase;
-use luya\admin\ngrest\render\RenderCrud;
+use luya\admin\ngrest\render\{RenderCrud,RenderCrudView};
 
 class RenderCrudTest extends AdminTestCase
 {
@@ -26,5 +26,25 @@ class RenderCrudTest extends AdminTestCase
         $this->assertSame([
             '<span class="foobar" label="foo" ng-href="#click"><i class="material-icons">extension</i><span> foo</span></span>',
         ], $crud->getSettingButtonDefinitions());
+    }
+
+    public function testSetView()
+    {
+        $crud = $this->getCrud();
+        
+        $crud->view = RenderCrudView::class;
+        $this->assertTrue(is_a($crud->view, RenderCrudView::class));
+        
+        $crud->view = ['class' => RenderCrudView::class];
+        $this->assertTrue(is_a($crud->view, RenderCrudView::class));
+        
+        $crud->view = new RenderCrudView();
+        $this->assertTrue(is_a($crud->view, RenderCrudView::class));
+    }
+
+    public function testGetView()
+    {
+        $crud = $this->getCrud();
+        $this->assertTrue(is_a($crud->view, RenderCrudView::class));
     }
 }